highlight double shifts
in order to adjust the highlight when players are moved, needed to feed back the data on every move.
because the smarts are all done in the python code.
this is ok though - since it now allows us to create the table in only one spot - and don't have it created in the template using django as well.